# Patch to fingerprint namedtuples.