@c200chromebook 2020-08-18
Under the hood is this how numpy structured data is represented? Would those of us using numpy dtypes get benefit?