@stuartarchibald 2020-08-18
> Under the hood is this how numpy structured data is represented? Would those of us using numpy dtypes get benefit?

This patch solely addresses instances of things inheriting from `tuple`, explicitly `namedtuple`, it won't impact anything else. There's fast path typecode lookups etc for common dtypes in the dispatcher already but not for structured dtypes.