@stuartarchibald 2020-08-25
TODO: benchmark this.