@EPronovost 2020-09-13
Nice work! I'd be curious how much of an improvement a previous PR of mine yields once this merges. https://github.com/numba/numba/pull/5840

One namedtuple-related question I've had for a while: Currently numba does not support subclasses of `pt.NamedTuple` which define methods.  For example,

```
class Point(typing.NamedTuple):
    x: int
    y: int
        
        
    def mag(self):
        return np.sqrt(self.x**2 + self.y**2)
```

Do you think it would be feasible to support these (maybe not in this PR)?