@sklam 2020-10-08
Needs decref: https://github.com/sklam/numba/commit/e984eb9406e37d71b8243562d0df849ac92087ac