@sklam 2020-10-08
Here's a microbenchmark:

```python
from numba import njit

from collections import namedtuple


@njit
def foo(x):
    return



for i in range(1, 8):
    n = 2**i
    print(f'n = {n}')
    tupcls = namedtuple("My", [f'f{x}' for x in range(n)])
    tup = tupcls(*range(n))
    foo(tup) #warmup
    %timeit foo(tup)
```
So, testing essentially the dispatch speed.

Results:

before PR:
```
n = 2
15.9 µs ± 540 ns per loop (mean ± std. dev. of 7 runs, 100000 loops each)
n = 4
21.1 µs ± 1.11 µs per loop (mean ± std. dev. of 7 runs, 10000 loops each)
n = 8
30.6 µs ± 637 ns per loop (mean ± std. dev. of 7 runs, 10000 loops each)
n = 16
48.7 µs ± 1.51 µs per loop (mean ± std. dev. of 7 runs, 10000 loops each)
n = 32
87.7 µs ± 2.31 µs per loop (mean ± std. dev. of 7 runs, 10000 loops each)
n = 64
164 µs ± 3.49 µs per loop (mean ± std. dev. of 7 runs, 10000 loops each)
n = 128
312 µs ± 8.48 µs per loop (mean ± std. dev. of 7 runs, 1000 loops each)
```

PR limit to namedtuple size 10
```
n = 2
865 ns ± 24 ns per loop (mean ± std. dev. of 7 runs, 1000000 loops each)
n = 4
1.08 µs ± 25.8 ns per loop (mean ± std. dev. of 7 runs, 1000000 loops each)
n = 8
1.34 µs ± 32.7 ns per loop (mean ± std. dev. of 7 runs, 1000000 loops each)
n = 16
51.9 µs ± 510 ns per loop (mean ± std. dev. of 7 runs, 10000 loops each)
n = 32
94.6 µs ± 3.23 µs per loop (mean ± std. dev. of 7 runs, 10000 loops each)
n = 64
178 µs ± 5.43 µs per loop (mean ± std. dev. of 7 runs, 10000 loops each)
n = 128
373 µs ± 19.2 µs per loop (mean ± std. dev. of 7 runs, 1000 loops each)
```

PR unlimited nametuple size
```
n = 2
841 ns ± 27.5 ns per loop (mean ± std. dev. of 7 runs, 1000000 loops each)
n = 4
986 ns ± 7.32 ns per loop (mean ± std. dev. of 7 runs, 1000000 loops each)
n = 8
1.35 µs ± 60 ns per loop (mean ± std. dev. of 7 runs, 1000000 loops each)
n = 16
2.06 µs ± 81.7 ns per loop (mean ± std. dev. of 7 runs, 100000 loops each)
n = 32
3.05 µs ± 132 ns per loop (mean ± std. dev. of 7 runs, 100000 loops each)
n = 64
5.41 µs ± 206 ns per loop (mean ± std. dev. of 7 runs, 100000 loops each)
n = 128
9.83 µs ± 413 ns per loop (mean ± std. dev. of 7 runs, 100000 loops each)
```

I think it is worth making it unlimited.