@stuartarchibald 2020-10-09
Thanks for the review @sklam, using the benchmark above I see this on mainline @ 9381206
```
n = 2
40.8 µs ± 1.13 µs per loop (mean ± std. dev. of 7 runs, 10000 loops each)
n = 4
49.3 µs ± 38.9 ns per loop (mean ± std. dev. of 7 runs, 10000 loops each)
n = 8
66.9 µs ± 1.36 µs per loop (mean ± std. dev. of 7 runs, 10000 loops each)
n = 16
99.3 µs ± 830 ns per loop (mean ± std. dev. of 7 runs, 10000 loops each)
n = 32
165 µs ± 224 ns per loop (mean ± std. dev. of 7 runs, 10000 loops each)
n = 64
296 µs ± 203 ns per loop (mean ± std. dev. of 7 runs, 1000 loops each)
n = 128
567 µs ± 9.38 µs per loop (mean ± std. dev. of 7 runs, 1000 loops each)
```
and this with this PR @ 32ae0f6
```
n = 2
1.32 µs ± 18.2 ns per loop (mean ± std. dev. of 7 runs, 1000000 loops each)
n = 4
1.56 µs ± 3.08 ns per loop (mean ± std. dev. of 7 runs, 1000000 loops each)
n = 8
2.02 µs ± 3.95 ns per loop (mean ± std. dev. of 7 runs, 1000000 loops each)
n = 16
3.08 µs ± 10.9 ns per loop (mean ± std. dev. of 7 runs, 100000 loops each)
n = 32
4.9 µs ± 3.75 ns per loop (mean ± std. dev. of 7 runs, 100000 loops each)
n = 64
8.8 µs ± 26.8 ns per loop (mean ± std. dev. of 7 runs, 100000 loops each)
n = 128
16.2 µs ± 23.3 ns per loop (mean ± std. dev. of 7 runs, 100000 loops each)
```
so this PR makes dispatch for `namedtuple` over an order of magnitude faster.