@stuartarchibald 2020-10-09
@EPronovost 

> One namedtuple-related question I've had for a while: Currently numba does not support subclasses of `pt.NamedTuple` which define methods. For example,
> 
> ```
> class Point(typing.NamedTuple):
>     x: int
>     y: int
>         
>         
>     def mag(self):
>         return np.sqrt(self.x**2 + self.y**2)
> ```
> 
> Do you think it would be feasible to support these (maybe not in this PR)?

I think it may be possible to support this at some point, but is out of scope for this PR, perhaps open a ticket with it as a feature request?