@stuartarchibald 2020-10-09
> Needs decrefs. See patch [sklam@e984eb9](https://github.com/sklam/numba/commit/e984eb9406e37d71b8243562d0df849ac92087ac)

Thanks, merged in.